# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Examance API",
    "description": "Privacy-first anonymous exam grading backend.\n\n## Key Security & Architecture Features\n- **Zero-Knowledge Encryption**: Student PII, submission scans, and annotations are encrypted client-side (Argon2id + AES-256-GCM) before reaching the server.\n- **Multi-factor sign-in**: a session needs a passkey (user verification required) or two distinct factors out of password, passkey and authenticator (TOTP). Fewer than two enrolled means an enrollment-scoped token that reaches nothing else.\n- **Recoverable data key**: the client's key is random and wrapped per factor (`key_envelopes`), so a password reset re-wraps it rather than orphaning every vault. The server holds ciphertext only.\n- **Session Hygiene**: State stored in HttpOnly cookies (`access_token` 15 min, `refresh_token` 7 days with rotation & reuse detection).\n- **Privacy Safeguards**: LaTeX requests redacted in logs (`LaTeXRequest`), IP addresses stored as SHA-256 hashes, k-anonymity (k >= 5) enforced on class stats.\n- **LaTeX Engine**: Compilation rendered using sandboxed Tectonic (`tectonic --untrusted`).",
    "version": "0.0.0-dev"
  },
  "paths": {
...
    "/api/v1/user/capabilities": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Capabilities",
        "description": "The account's storage mode (null until chosen) and what it may use.",
        "operationId": "get_capabilities_api_v1_user_capabilities_get",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/storage-mode": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Set Storage Mode",
        "description": "Records the account's storage mode, compare-and-set on `expected`. Moving the results\nbetween server and browser is the client's job and happens before this call; this only\nstates where they now live.",
        "operationId": "set_storage_mode_api_v1_user_storage_mode_put",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageModeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/purge-server-student-data": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Purge Server Student Data",
        "description": "Soft-delete all student identities and scan submissions belonging to the current teacher\nwith a 7-day retention grace period before hard deletion.\n\nLaTeX exercise templates and exam structures remain intact \u2014 including the\nresource files attached to exercises, which are teacher-authored content,\nnot student data. A file that does contain personal data (e.g. a photo of a\npupil) has to be deleted with the exercise or the file itself.",
        "operationId": "purge_server_student_data_api_v1_user_purge_server_student_data_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Purge Server Student Data Api V1 User Purge Server Student Data Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/restore-server-data": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Restore Server Data",
        "description": "Restore soft-deleted student identities and scan submissions for the current teacher\nif they are within the 7-day retention grace period.",
        "operationId": "restore_server_data_api_v1_user_restore_server_data_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Restore Server Data Api V1 User Restore Server Data Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/me/export": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Export Own Data",
        "description": "GDPR Art. 15/20 \u2014 machine-readable copy of the account holder's own data.\n\nCovers only what the server holds *about the teacher*: account fields, the\nexams they authored, and their audit trail. Student payloads are deliberately\nexcluded \u2014 they are encrypted with a key the server never has, and they are\nnot the teacher's personal data. Use the client-side per-student export for\na student's Art. 15 request.",
        "operationId": "export_own_data_api_v1_user_me_export_get",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Own Data Api V1 User Me Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/me": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete Own Account",
        "description": "GDPR Art. 17 \u2014 erase the account holder's own account and authored content.\n\nStudent identities and submissions under the teacher's exams are soft-deleted\nwith the standard grace period and erased by the retention job, matching\n`purge-server-student-data`.\n\nAudit rows are kept, with `teacher_id` nulled by the FK's ON DELETE SET NULL\nand the email snapshot left in place: Art. 17(3)(b) permits retaining what is\nneeded for a legal obligation, and the trail exists to evidence lawful\nhandling of student data. Those rows age out under AUDIT_LOG_RETENTION_DAYS\nrather than living forever.",
        "operationId": "delete_own_account_api_v1_user_me_delete",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Own Account Api V1 User Me Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapabilitiesOut": {
        "properties": {
          "storage_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "all-server",
                  "hybrid"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Mode"
          },
          "allowed_storage_modes": {
            "items": {
              "type": "string",
              "enum": [
                "all-server",
                "hybrid"
              ]
            },
            "type": "array",
            "title": "Allowed Storage Modes"
          },
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "storage_mode",
          "allowed_storage_modes",
          "features"
        ],
        "title": "CapabilitiesOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StorageModeUpdate": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "all-server",
              "hybrid"
            ],
            "title": "Mode"
          },
          "expected": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "all-server",
                  "hybrid"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "expected"
        ],
        "title": "StorageModeUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}